Local Cache for User Repo
The current user is being requested on every single request to the server. This is a relatively quick postgres db hit, but after adding in latency, it can make a bit of a difference in overall response time. Because users are relatively small objects, I added a generic repo cache around it that will keep all single user requests cached for 10 minutes. If I ever get a large amount of users I'll have to look at optimizing the memory space here, but for now this makes a noticeable difference.